@@ -1,4 +1,4 @@
-from typing import Callable, Dict, List, Optional, Union
+from typing import Callable, Optional, Union
 
 from assertionengine.formatter import FormatRules
 from assertionengine.formatter import Formatter as ASFormatter
@@ -24,7 +24,7 @@ class Formatter(ASFormatter, LibraryComponent):
         keyword: Optional[FormatterKeywords] = None,
         *formatters: Union[FormatingRules, LambdaFunction],
         scope: Scope = Scope.Global,
-    ) -> Dict[str, List[str]]:
+    ) -> dict[str, list[str]]:
         """Set keyword assertion formatter with defined scope.
 
         Returns the old formatter from the scope.
@@ -97,7 +97,7 @@ class Formatter(ASFormatter, LibraryComponent):
     @keyword_deco(tags=("Config",))
     def set_assertion_formatters(
         self, formatters: FormatterTypes, scope: Scope = Scope.Suite
-    ) -> Dict[str, List[str]]:
+    ) -> dict[str, list[str]]:
         """Set keywords formatters for assertions.
 
         | =Arguments= | =Description= |
