@@ -19,7 +19,7 @@ from os import PathLike
 from pathlib import Path
 from sys import maxsize
 from time import sleep
-from typing import Any, Dict, List, Optional
+from typing import Any, Optional
 
 from robot.api.deco import keyword
 from robot.utils import DotDict
@@ -87,8 +87,8 @@ class Promises(LibraryComponent):
         return kw.lower().replace(" ", "").replace("_", "")
 
     def resolve_arguments(self, kw: str, *args):
-        positional: List[Any] = []
-        named: Dict[str, Any] = {}
+        positional: list[Any] = []
+        named: dict[str, Any] = {}
         logger.debug(f"*args {args}")
 
         arg_names = []
