@@ -16,7 +16,7 @@ import json
 import re
 import time
 from datetime import timedelta
-from typing import Any, Dict, Optional, Union
+from typing import Any, Optional, Union
 
 from robot.libraries.BuiltIn import BuiltIn
 
@@ -140,7 +140,7 @@ class Waiter(LibraryComponent):
     ):
         selector = self.resolve_selector(selector)
         with self.playwright.grpc_channel() as stub:
-            options: Dict[str, object] = {"state": state.name}
+            options: dict[str, object] = {"state": state.name}
             if timeout:
                 options["timeout"] = self.get_timeout(timeout)
             options_json = json.dumps(options)
@@ -210,7 +210,7 @@ class Waiter(LibraryComponent):
     ):
         selector = self.resolve_selector(selector)
         with self.playwright.grpc_channel() as stub:
-            options: Dict[str, int] = {}
+            options: dict[str, int] = {}
             if polling != "raf":
                 options["polling"] = self.convert_timeout(polling)  # type: ignore
             if timeout:
