@@ -12,7 +12,7 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 import inspect
-from typing import Any, Callable, Dict, Tuple
+from typing import Any, Callable
 
 from Browser.utils import logger
 
@@ -46,8 +46,8 @@ def _is_deprecated_attribute(method: Callable, deprecated_arg, args, kwargs):
 
 
 def convert_pos_args_to_named(
-    deprecated_pos_args: Tuple[Any, ...],
-    old_args: Dict[str, Any],
+    deprecated_pos_args: tuple[Any, ...],
+    old_args: dict[str, Any],
     keyword_name: str,
     additional_msg: str = "",
 ):
