@@ -12,11 +12,11 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 from enum import Enum
-from typing import Any, Dict, List, TypeVar
+from typing import Any, TypeVar
 
 
-def locals_to_params(args: Dict) -> Dict:
-    copy: Dict[str, Any] = {}
+def locals_to_params(args: dict) -> dict:
+    copy: dict[str, Any] = {}
     for key in args:
         if key == "self":
             continue
@@ -37,7 +37,7 @@ def locals_to_params(args: Dict) -> Dict:
 T = TypeVar("T")
 
 
-def find_by_id(_id: str, item_list: List[Dict[str, T]], log_error=True) -> Dict[str, T]:
+def find_by_id(_id: str, item_list: list[dict[str, T]], log_error=True) -> dict[str, T]:
     from ..utils import logger
 
     def filter_fn(item):
