@@ -19,7 +19,7 @@ import socket
 import string
 import subprocess
 from pathlib import Path
-from typing import Any, Tuple, Union
+from typing import Any, Union
 
 from robot.libraries.BuiltIn import BuiltIn
 
@@ -35,7 +35,7 @@ def find_free_port() -> int:
         return s.getsockname()[1]
 
 
-def spawn_node_process(output_dir: Path) -> Tuple[subprocess.Popen, str]:
+def spawn_node_process(output_dir: Path) -> tuple[subprocess.Popen, str]:
     """
     Spawn an rfbrowser node process, that can be shared between library instances.
 
@@ -87,7 +87,7 @@ def get_normalized_keyword(keyword: str) -> str:
     return new_keyword.lower().replace(" ", "_")
 
 
-def keyword(name: Any = None, tags: Tuple = (), types: Tuple = ()):
+def keyword(name: Any = None, tags: tuple = (), types: tuple = ()):
     if inspect.isroutine(name):
         return keyword()(name)
 
